# Conflicts:
#	src/main/java/edu/metrostate/Show.java
#	src/main/java/gui/seasonoverview/SeasonOverviewController.java
#	src/main/java/gui/showoverview/ShowOverviewController.java
#	src/main/resources/fxml/showoverview/ShowOverview.fxml

+++ src/main/java/edu/metrostate/Reviewable.java
@@ -3,7 +3,7 @@ package edu.metrostate;
 public interface Reviewable {
     //"public" declaration is not needed as methods in interfaces are understood to be public
 
-    int getStars();
+    double getStars();
     int getShowId();
     void recalculateStars(Review review);
 }
